Part 1 of the compiler warning patch. Will update wv module shortly...
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@3312 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fl_BlockLayout.cpp
@@ -4452,13 +4452,13 @@ void    fl_BlockLayout::StartList( const XML_Char * style)
 		else 
 			startv = 1;
 		if (szAlign)
-			fAlign = atof(szAlign);
+			fAlign = (float)atof(szAlign);
 		else
-			fAlign = 0.25;
+			fAlign = (float)0.25;
 		if (szIndent)
-			fIndent = atof(szIndent);
+			fIndent = (float)atof(szIndent);
 		else
-			fIndent = -0.25;
+			fIndent = (float)-0.25;
 	}
 	else
 	{
